requirements.txt
streamlit
pandas
plotly
python-dotenv
huggingface_hub
requests
textblob